requirements.txt
streamlit
crewai
crewai-tools
langchain-google-genai
langchain-community
fpdf2
tavily-python